regex to stable
