+++ src/bin/libs/mod.rs
@@ -1,3 +1,4 @@
 pub mod loops;
 pub mod io;
+pub mod regex;
 mod ping;
